pyproject.toml
[project]
name = "tenko"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "brotli>=1.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "textual>=6.6.0",
    "urllib3>=2.5.0",
]
